removed diagnostics/changed CAN client to publisher. renamed files to accurately represent function
