+++ src/moa/moa_controllers/package.xml
@@ -15,7 +15,8 @@
   <exec_depend>rclpy</exec_depend>
   <exec_depend>ackermann_msgs</exec_depend>
   <exec_depend>moa_msgs</exec_depend>
-  <exec_depend>diagnostics</exec_depend>
+  <exec_depend>std_msgs</exec_depend>
+  <exec_depend>builtin_interfaces</exec_depend>
 
 
   <export>
